chatbot/app/requirements.txt
fastapi
uvicorn
openai
python-dotenv
prometheus-client
